requirements.txt
numpy 
pandas 
fastapi
pillow  
matplotlib
torchinfo
uvicorn[standard]
python-multipart